API modeler - 1.1.b : for mocks, move methods definition to .h files
