+++ src/tests/src/solver/modeler/api/LinearProblemBuilder.cmake
@@ -1,9 +1,5 @@
-set(FILES_TO_LINK
-        mock-fillers/OneVarFiller.cpp
-        mock-fillers/OneConstraintFiller.cpp
-        mock-fillers/TwoVarsTwoConstraints.cpp)
 
-add_executable(test-modeler-LP-builder testModelerLPbuilder.cpp ${FILES_TO_LINK})
+add_executable(test-modeler-LP-builder testModelerLPbuilder.cpp)
 
 target_include_directories(test-modeler-LP-builder
         PRIVATE
