--- src/tests/src/solver/modeler/api/mock-fillers/OneConstraintFiller.cpp
@@ -1,19 +1,0 @@
-#include "OneConstraintFiller.h"
-
-namespace Antares::Solver::Modeler::Api
-{
-
-void OneConstraintFiller::addVariables()
-{
-}
-
-void OneConstraintFiller::addConstraints()
-{
-    linearProblem_->addConstraint(1, 2, "constraint-by-OneConstraintFiller");
-}
-
-void OneConstraintFiller::addObjective()
-{
-}
-
-} // namespace Antares::Solver::Modeler::Api
