+++ src/tests/src/solver/modeler/api/mock-fillers/OneConstraintFiller.h
@@ -15,4 +15,17 @@ public:
     void addObjective() override;
 };
 
+void OneConstraintFiller::addVariables()
+{
+}
+
+void OneConstraintFiller::addConstraints()
+{
+    linearProblem_->addConstraint(1, 2, "constraint-by-OneConstraintFiller");
+}
+
+void OneConstraintFiller::addObjective()
+{
+}
+
 } // namespace Antares::Solver::Modeler::Api
