+++ src/tests/src/solver/modeler/api/mock-fillers/OneVarFiller.h
@@ -18,4 +18,19 @@ private:
     std::string added_var_name_ = "var-by-OneVarFiller";
 };
 
+void OneVarFiller::addVariables()
+{
+    linearProblem_->addNumVariable(0, 1, added_var_name_);
+}
+
+void OneVarFiller::addConstraints()
+{
+}
+
+void OneVarFiller::addObjective()
+{
+    auto* var = linearProblem_->getVariable(added_var_name_);
+    linearProblem_->setObjectiveCoefficient(var, 1);
+}
+
 } // namespace Antares::Solver::Modeler::Api
